refactor: env.py -> environment.py (match the Environment class; last abbreviation)
