+++ libs/agno/agno/environments/__init__.py
@@ -10,7 +10,7 @@ without importing scorer.
 """
 
 from agno.environments._engine import AttemptResult, StopReason
-from agno.environments.env import Environment, Task
+from agno.environments.environment import Environment, Task
 from agno.environments.exporters import ExportReport, ato_sft_jsonl, to_sft_jsonl
 from agno.environments.runner import EnvironmentDiff, EnvironmentRunResult, TaskResult, arun_rollouts, run_rollouts
 from agno.scorer import FingerprintError, MismatchError
